backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.8",
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.44.0",
]
